do hot-reloading of world parameters

+++ world/pathfinding.go
@@ -80,7 +80,7 @@ func (p *Pathfinding) Initialize(m Matrix) {
 func (p *Pathfinding) computePath(parents []int, end int) (path []Pt) {
 	node := end
 	for node >= 0 {
-		path = append(path, p.m.IndexToPt(I(int64(node))))
+		path = append(path, p.m.IndexToPt(I(node)))
 		node = parents[node]
 	}
 	slices.Reverse(path)
